+++ src/p_012.rs
@@ -1,6 +1,6 @@
 #[allow(dead_code)]
-pub fn problem_012() -> u64 {
-    let mut n: u64 = 1;
+pub fn problem_012() -> i128 {
+    let mut n: i128 = 1;
     let mut incrementer = 2;
     loop {
         n += incrementer;
@@ -10,10 +10,10 @@ pub fn problem_012() -> u64 {
             continue;
         }
 
-        let n_sqrt = (n as f64).sqrt() as u64;
+        let n_sqrt = (n as f64).sqrt() as i128;
         let count = (2..=n_sqrt)
             .map(|i| if n % i == 0 { 2 } else { 0 })
-            .sum::<u64>()
+            .sum::<i128>()
             + 2;
         if count > 500 {
             return n;
