+++ src/p_015.rs
@@ -1,4 +1,4 @@
-fn comb(n: u64, k: u64) -> u64 {
+fn comb(n: i128, k: i128) -> i128 {
     if k == 0 {
         1
     } else if k > n / 2 {
@@ -9,7 +9,7 @@ fn comb(n: u64, k: u64) -> u64 {
 }
 
 #[allow(dead_code)]
-pub fn problem_015() -> u64 {
+pub fn problem_015() -> i128 {
     let k = 20;
     let n = k * 2;
     comb(n, k)
