+++ src/p_017.rs
@@ -1,4 +1,4 @@
 #[allow(dead_code)]
-pub fn problem_017() -> u64 {
+pub fn problem_017() -> i128 {
     0
 }
